+++ tools/ark_frequent_instructions.py
@@ -33,7 +33,9 @@ super_insts = [
     "NEQ_CONST_JUMP_IF_TRUE",
     "CALL_SYMBOL",
     "GET_FIELD_FROM_SYMBOL",
-    "GET_FIELD_FROM_SYMBOL_INDEX"
+    "GET_FIELD_FROM_SYMBOL_INDEX",
+    "AT_SYM_SYM",
+    "AT_SYM_INDEX_SYM_INDEX"
 ]
 
 executable = None